+++ sdk/examples/wallet/nft_collection/00_mint_issuer_nft.rs
@@ -42,14 +42,10 @@ async fn main() -> Result<()> {
 
     // Issue the minting transaction and wait for it's inclusion
     println!("Sending NFT minting transaction...");
-    let nft_mint_params = vec![MintNftParams {
-        address: None,
-        immutable_metadata: Some(b"This NFT will be the issuer from the awesome NFT collection".to_vec()),
-        issuer: None,
-        metadata: None,
-        sender: None,
-        tag: None,
-    }];
+    let nft_mint_params = vec![
+        MintNftParams::new()
+            .with_immutable_metadata(b"This NFT will be the issuer from the awesome NFT collection".to_vec()),
+    ];
     let transaction = account.mint_nfts(nft_mint_params, None).await?;
     wait_for_inclusion(&transaction.transaction_id, &account).await?;